requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
attrs==24.2.0
discord==2.3.2
discord.py==2.4.0
enum-tools==0.12.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
psycopg2-binary==2.9.10
Pygments==2.18.0
python-dateutil==2.9.0.post0
six==1.16.0
twitchAPI==4.3.1
typing_extensions==4.12.2
yarl==1.18.0
